* 'master' of github.com:tensorflow/tensorflow: (177 commits)
  Fix ViewOp to have at most one offset operand
  AffineLoopFusion: Prevent fusion of multi-out-edge producer loops
  Extend conversion of SubViewOp to llvm to also support cases where size and stride are constant (i.e., there are no size and stride operands).
  [spirv] Add spv.SubgroupBallotKHROp
  [Linalg] Update/fix documentation for linalg.indexed_generic.
  Add replay_computation target for mlir_gpu backend.
  compat: Update forward compatibility horizon to 2019-12-03
  Add deprecation note for GL delegate.
  Add linkage support to LLVMFuncOp
  Combine all HLO dialect exporter tests in a single file.
  Only warn about unoptimized datasets every 30s
  Avoid heap allocation of ExecuteNodeArgs
  change optimized quantized int8 path 4d mean to be integer-only.
  Make the TFLite with select TF ops for iOS composable
  [XLA] Refactor Executable::ExecuteAsyncOnStream.
  Fix tensor initialization.
  Make shape_inference_helper buildable for Android.
  Handle 32 bit integer paddings operand for tf.PadV2 lowering
  Remove dependence on core/lib/core/blocking_counter.
  Throw an explicit error if user call TPUStrategy experimental_run_v2 in eager mode with a python function.
  ...

+++ tensorflow/lite/tools/benchmark/benchmark_tflite_model.cc
@@ -492,13 +492,15 @@ TfLiteStatus BenchmarkTfLiteModel::PrepareInputData() {
     } else if (t->type == kTfLiteUInt8) {
       int low = has_value_range ? low_range : 0;
       int high = has_value_range ? high_range : 254;
+      // std::uniform_int_distribution is specified not to support char types.
       t_data = CreateInputTensorData<uint8_t>(
-          num_elements, std::uniform_int_distribution<uint8_t>(low, high));
+          num_elements, std::uniform_int_distribution<uint32_t>(low, high));
     } else if (t->type == kTfLiteInt8) {
       int low = has_value_range ? low_range : -127;
       int high = has_value_range ? high_range : 127;
+      // std::uniform_int_distribution is specified not to support char types.
       t_data = CreateInputTensorData<int8_t>(
-          num_elements, std::uniform_int_distribution<int8_t>(low, high));
+          num_elements, std::uniform_int_distribution<int32_t>(low, high));
     } else if (t->type == kTfLiteString) {
       // TODO(haoliang): No need to cache string tensors right now.
     } else {
